requirements.txt
streamlit
pytesseract
Pillow
openpyxl
